fix(integration-jaas): fix JAAS tests with kubernetes_cloud by avoiding to create a model
We can't create a model in JAAS tests because we can't remove a cloud if it is used by a model. In
Juju if you don't specify a cloud, it picks the controller's cloud, but JAAS doesn't and returns an
error if multiple clouds are available, which is the case because we don't remove the clouds created
in the tests.